Comment out "checkOwners" since it doesn't work expectedly now.
It's possible to override "image" with any docker images with violation because now the validation admission webhook for Deployment doesn't check "UPDATE" operation, then Kritis allows any docker images in ReplicaSet and Pod creation because of the "checkOwners".

If we add "UPDATE" operation to the webhook for Deployment, the webhook can get AdmissionReview for Deployment update but it cause another issue which we can't know the admission comes from "UPDATE" or "DELETE" operation. (Both sends "UPDATE" admission)
It's probably because the admission's Kind is "extensions/v1beta1" and it doesn't support "DELETE" operation.

So for now it's better to disable the "checkOwners" so that we can reject docker images when Deployment updated.